test project script
